rename packet_deserializer to packet_deserializer_stage
